Dernier code non fonctionnel pour curve height avec stand by
