+++ PolygonCSV/polygon_fillet.csv
@@ -1,8 +1,8 @@
-0,20,0
-0,180,-5.858
-20,200,0
-180,200,-5.858
-200,180,0
-200,20,-5.858
-180,0,0
-20,0,-5.858
+20,0,0
+180,0,5.858
+200,20,0
+200,180,5.858
+180,200,0
+20,200,5.858
+0,180,0
+0,20,5.858
