+++ PolygonCSV/polygon_fillet2.csv
@@ -1,0 +1,9 @@
+10,10,0
+1,10,0.293
+0,9,0
+0,1,0.293
+1,0,0
+27.586,0,0.617
+28.293,1.707,0
+11.707,18.293,0.617
+10,17.586,0
