+++ PolygonCSV/polygon_fillet3.csv
@@ -1,0 +1,6 @@
+0,7.586,0
+0,1,0.293
+1,0,0
+7.586,0,0.617
+8.283,1.707,0
+1.707,8.293,0.617
